requirements.txt
asgiref==3.3.1
colorama==0.4.4
Django==3.1.5
django-ckeditor==6.0.0
django-crispy-forms==1.10.0
django-debug-toolbar==3.2
django-filter==2.4.0
django-js-asset==1.2.2
djangorestframework==3.12.2
et-xmlfile==1.0.1
jdcal==1.4.1
Markdown==3.3.3
openpyxl==3.0.6
phonenumbers==8.12.15
Pillow==8.1.0
python-memcached==1.59
pytz==2020.5
qrcode==6.1
six==1.15.0
sqlparse==0.4.1
xlwt==1.3.0
